Pizza Ordering script
Important notes;

I have no idea if this will work for all countries, I haven't tested it yet but in theory it should work. No idea how long it takes for the confirm message to come through. People in the media who've tried it before have said the service often fails so..